backend/requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.2
django-rest-authtoken==2.1.4
djangorestframework==3.14.0
pytz==2023.3
sqlparse==0.4.4
psycopg2-binary
django-cors-headers
djangorestframework-simplejwt==5.2.2
pillow==9.5.0
